requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
jinja2==3.1.6
python-dotenv==1.2.2
google-genai==1.72.0
pydantic==2.12.5
groq==1.1.2
